+++ sv-parser-pp/testcases/test15.sv
@@ -1,0 +1,4 @@
+`define MOD_INST u_mysubmod
+module mymod;
+mysubmod `MOD_INST ();
+endmodule