+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/sv-parser/compare/v0.8.2...Unreleased) - ReleaseDate
 
+* [Fixed] textmacro without arguments issue [#20](https://github.com/dalance/sv-parser/issues/20)
+
 ## [v0.8.2](https://github.com/dalance/sv-parser/compare/v0.8.1...v0.8.2) - 2020-08-28
 
 * [Fixed] empty ifdef issue [#10](https://github.com/dalance/sv-parser/issues/10)